+++ crates/store/src/server/rpc_api.rs
@@ -1,4 +1,6 @@
 use miden_node_proto::convert;
+use miden_node_proto::domain::block::InvalidBlockRange;
+use miden_node_proto::errors::MissingFieldHelper;
 use miden_node_proto::generated::store::rpc_server;
 use miden_node_proto::generated::{self as proto};
 use miden_node_utils::limiter::{
@@ -10,6 +12,7 @@ use miden_node_utils::limiter::{
 };
 use miden_protocol::Word;
 use miden_protocol::account::AccountId;
+use miden_protocol::block::BlockNumber;
 use miden_protocol::note::NoteId;
 use tonic::{Request, Response, Status};
 use tracing::{debug, info};
@@ -24,6 +27,7 @@ use crate::errors::{
     NoteSyncError,
     SyncAccountStorageMapsError,
     SyncAccountVaultError,
+    SyncChainMmrError,
     SyncNullifiersError,
     SyncTransactionsError,
 };
@@ -118,54 +122,6 @@ impl rpc_server::Rpc for StoreApi {
         }))
     }
 
-    /// Returns info which can be used by the client to sync up to the latest state of the chain
-    /// for the objects the client is interested in.
-    async fn sync_state(
-        &self,
-        request: Request<proto::rpc::SyncStateRequest>,
-    ) -> Result<Response<proto::rpc::SyncStateResponse>, Status> {
-        let request = request.into_inner();
-
-        let account_ids: Vec<AccountId> = read_account_ids::<Status>(&request.account_ids)?;
-
-        let (state, delta) = self
-            .state
-            .sync_state(request.block_num.into(), account_ids, request.note_tags)
-            .await
-            .map_err(internal_error)?;
-
-        let accounts = state
-            .account_updates
-            .into_iter()
-            .map(|account_info| proto::account::AccountSummary {
-                account_id: Some(account_info.account_id.into()),
-                account_commitment: Some(account_info.account_commitment.into()),
-                block_num: account_info.block_num.as_u32(),
-            })
-            .collect();
-
-        let transactions = state
-            .transactions
-            .into_iter()
-            .map(|transaction_summary| proto::transaction::TransactionSummary {
-                account_id: Some(transaction_summary.account_id.into()),
-                block_num: transaction_summary.block_num.as_u32(),
-                transaction_id: Some(transaction_summary.transaction_id.into()),
-            })
-            .collect();
-
-        let notes = state.notes.into_iter().map(Into::into).collect();
-
-        Ok(Response::new(proto::rpc::SyncStateResponse {
-            chain_tip: self.state.latest_block_num().await.as_u32(),
-            block_header: Some(state.block_header.into()),
-            mmr_delta: Some(delta.into()),
-            accounts,
-            transactions,
-            notes,
-        }))
-    }
-
     /// Returns info which can be used by the client to sync note state.
     async fn sync_notes(
         &self,
@@ -194,6 +150,58 @@ impl rpc_server::Rpc for StoreApi {
             block_header: Some(state.block_header.into()),
             mmr_path: Some(mmr_proof.merkle_path.into()),
             notes,
+        }))
+    }
+
+    /// Returns chain MMR updates within a block range.
+    async fn sync_chain_mmr(
+        &self,
+        request: Request<proto::rpc::SyncChainMmrRequest>,
+    ) -> Result<Response<proto::rpc::SyncChainMmrResponse>, Status> {
+        // TODO find a reasonable upper boundary
+        const MAX_BLOCKS: u32 = 1 << 20;
+
+        let request = request.into_inner();
+        let chain_tip = self.state.latest_block_num().await;
+
+        let block_range = request
+            .block_range
+            .ok_or_else(|| proto::rpc::SyncChainMmrRequest::missing_field(stringify!(block_range)))
+            .map_err(SyncChainMmrError::DeserializationFailed)?;
+
+        let block_from = BlockNumber::from(block_range.block_from);
+        if block_from > chain_tip {
+            Err(SyncChainMmrError::FutureBlock { chain_tip, block_from })?;
+        }
+
+        let block_to = block_range.block_to.map_or(chain_tip, BlockNumber::from).min(chain_tip);
+
+        if block_from > block_to {
+            Err(SyncChainMmrError::InvalidBlockRange(InvalidBlockRange::StartGreaterThanEnd {
+                start: block_from,
+                end: block_to,
+            }))?;
+        }
+        let block_range = block_from..=block_to;
+        let len = 1 + block_range.end().as_u32() - block_range.start().as_u32();
+        let trimmed_block_range = if len > MAX_BLOCKS {
+            block_from..=BlockNumber::from(block_from.as_u32() + MAX_BLOCKS)
+        } else {
+            block_range
+        };
+
+        let mmr_delta = self
+            .state
+            .sync_chain_mmr(trimmed_block_range.clone())
+            .await
+            .map_err(internal_error)?;
+
+        Ok(Response::new(proto::rpc::SyncChainMmrResponse {
+            pagination_info: Some(proto::rpc::PaginationInfo {
+                chain_tip: chain_tip.as_u32(),
+                block_num: trimmed_block_range.end().as_u32(),
+            }),
+            mmr_delta: Some(mmr_delta.into()),
         }))
     }